delete PAIRED intermediate files in SRR folder

+++ rnannot/RNAseq_annotate.py
@@ -552,8 +552,16 @@ if __name__ == '__main__':
             os.remove(hs2_file)
         # remove fastq files in SRR subdirectories
         for run in runs:
-            os.remove(path.join(args.outdir, args.name, run, 'output.fastq'))
-            os.remove(path.join(args.outdir, args.name, run, 'normalized.fastq'))
+            if path.exists(path.join(args.outdir, args.name, run, 'output.fastq')):
+                os.remove(path.join(args.outdir, args.name, run, 'output.fastq'))
+                os.remove(path.join(args.outdir, args.name, run, 'normalized.fastq'))
+            if path.exists(path.join(args.outdir, args.name, run, 'output_1.fastq')):
+                os.remove(path.join(args.outdir, args.name, run, 'output_1.fastq'))
+                os.remove(path.join(args.outdir, args.name, run, 'output_1_un.fastq')
+                os.remove(path.join(args.outdir, args.name, run, 'normalized_1.fastq'))
+                os.remove(path.join(args.outdir, args.name, run, 'output_2.fastq'))
+                os.remove(path.join(args.outdir, args.name, run, 'output_2_un.fastq'))
+                os.remove(path.join(args.outdir, args.name, run, 'normalized_2.fastq'))
             if path.exists(path.join(args.outdir, args.name, run, run + '_1.fastq')):
                 os.remove(path.join(args.outdir, args.name, run, run + '_1.fastq'))
                 os.remove(path.join(args.outdir, args.name, run, run + '_1_fastqc.zip'))
